エディタースクリプトを修正
