+++ Runtime/jp.ootr.WeatherWidget/Scripts/11_BaseClass.cs
@@ -7,5 +7,6 @@ namespace jp.ootr.WeatherWidget
     {
         [SerializeField] protected Animator animator;
         [SerializeField] public Image splashImage;
+        [SerializeField] public Sprite splashSprite;
     }
 }
